Rename file number_of_age
